refactor: UI 컴포넌트 FSD 규칙에 맞게 분리 및 배럴 구성
- src/components/* → src/shared/ui/* 이동 및 기존 디렉터리 제거
- shared/ui/index.ts를 배럴 전용(export-only)으로 전환
- 컴포넌트 분리: Button, Input, Card(+Header/Title/Content), Textarea, Select, Dialog, Table
- 로컬 폴더 구조 도입: shared/ui/{button,card,input,textarea,select,dialog,table}/
- 최상위 배럴(@shared/ui)에서 재-export 구성
- 사용처 import를 @shared/ui로 교체 (pages/PostsManagerPage.tsx, App.tsx 등)
- Header/Footer를 shared/ui로 이동하고 배럴에 추가
- 경로 alias(@shared/ui) 유지 및 타입 정리
