+++ src/App.tsx
@@ -1,8 +1,6 @@
+import PostsManagerPage from "@pages/PostsManagerPage"
+import { Footer, Header } from "@shared/ui"
 import { BrowserRouter as Router } from "react-router-dom"
-
-import Footer from "./components/Footer.tsx"
-import Header from "./components/Header.tsx"
-import PostsManagerPage from "./pages/PostsManagerPage.tsx"
 
 const App = () => {
   return (
