+++ src/pages/PostsManagerPage.tsx
@@ -1,7 +1,3 @@
-import { Edit2, MessageSquare, Plus, Search, ThumbsDown, ThumbsUp, Trash2 } from "lucide-react"
-import { useEffect, useState } from "react"
-import { useLocation, useNavigate } from "react-router-dom"
-
 import {
   Button,
   Card,
@@ -25,7 +21,10 @@ import {
   TableHeader,
   TableRow,
   Textarea,
-} from "../components"
+} from "@shared/ui"
+import { Edit2, MessageSquare, Plus, Search, ThumbsDown, ThumbsUp, Trash2 } from "lucide-react"
+import { useEffect, useState } from "react"
+import { useLocation, useNavigate } from "react-router-dom"
 
 const PostsManager = () => {
   const navigate = useNavigate()
