+++ src/shared/ui/dialog/index.ts
@@ -1,0 +1,4 @@
+export * from "@shared/ui/dialog/Dialog"
+export * from "@shared/ui/dialog/DialogTrigger"
+export * from "@shared/ui/dialog/DialogPortal"
+export * from "@shared/ui/dialog/DialogOverlay"
