+++ src/shared/ui/index.ts
@@ -1,0 +1,16 @@
+export { Button } from "@shared/ui/button"
+export { Card, CardHeader, CardTitle, CardContent } from "@shared/ui/card"
+export { Input } from "@shared/ui/input"
+export { Select, SelectGroup, SelectValue, SelectTrigger, SelectContent, SelectItem } from "@shared/ui/select"
+export { Textarea } from "@shared/ui/textarea"
+export {
+  Dialog,
+  DialogContent,
+  DialogHeader,
+  DialogOverlay,
+  DialogPortal,
+  DialogTitle,
+  DialogTrigger,
+} from "@shared/ui/dialog"
+export { Table, TableHeader, TableBody, TableRow, TableHead, TableCell } from "@shared/ui/table"
+export { Footer, Header } from "@shared/ui"
