+++ src/shared/ui/table/index.ts
@@ -1,0 +1,1 @@
+export * from "@shared/ui/table/Table";
